Build the scanner and utility code on Windows
Paperman only builds on Unix: it links against libsane and libdl,
polls the SANE select descriptor with poll(), looks the user up with
getpwuid() and getgrnam(), sets file groups with chown() and reads
passwords with termios. None of these exist in a MinGW build, so the
program cannot be compiled for Windows at all.

Provide a win32/ directory holding the SANE headers (which are in the
public domain) along with a stub implementation of the entry points
paperman uses. The stub reports no devices and refuses every
operation, so the built-in simulated scanner still works while real
scanners are left for a future libsane port. A tiny poll() shim covers
the SANE select loop, which is never entered without a device.

Guard the remaining POSIX calls: use GetUserName() for the user name,
skip the group-sharing chmod/chown (there are no Unix groups), drop the
dlsym() lookup of the optional sane_read_dup() extension, replace
setenv() with qputenv() and toggle console echo with SetConsoleMode()
when reading the server password. Also replace the Qt 3 Q_WS_WIN code
in util_getUsername() which no longer compiles with any supported Qt.

The .pro files gain a win32 block which uses pkg-config to find poppler
and podofo under MSYS2 and keeps the executable next to the sources so
the tests can find their fixtures.

+++ filemax.cpp
@@ -3459,7 +3459,7 @@ static void TIFFFlushData1 (TIFF *tiff)
 
 #define  is2DEncoding(sp) \
    (sp->b.groupoptions & GROUP3OPT_2DENCODING)
-#define  isAligned(p,t) ((((u_long)(p)) & (sizeof (t)-1)) == 0)
+#define  isAligned(p,t) ((((uintptr_t)(p)) & (sizeof (t)-1)) == 0)
 #define  TIFFroundup(x, y) (TIFFhowmany(x,y)*((uint32_t)(y)))
 #define  TIFFhowmany(x, y) ((((uint32_t)(x))+(((uint32_t)(y))-1))/((uint32_t)(y)))
 
@@ -4574,7 +4574,7 @@ static int add_image_header (chunk_info &chunk, byte *buf)
    ptr [12] = chunk.parts.size ();
 
    iptr = (unsigned *)(buf + POS_part0);
-   assert (!((long int)iptr & 3));
+   assert (!((uintptr_t)iptr & 3));
    for (i = 0; i < chunk.parts.size (); i++, iptr += 2)
       {
       part_info &part = chunk.parts [i];
